Disconnect GPIOs from GND
GPIOs are supposed to power up in floating input mode, so I tied some of
them to GND for easier routing. For some reason the input voltage sagged
to ~0.6 V until I disconnected these pins from the PCB though. I'm not
sure if this was the cause or I just messed up soldering something, but
disconnecting these just to be safe.