Add code and 3D files, update README and LICENSE
Added an Arduino sketch and updated the README with instructions for
building and flashing it.

Added 3D model files for the enclosure and updated the README with
instructions for assembling it.

Added a license for the Arduino sketch. Removed license information for
footprints that were copied from another project but have been removed
and were not used.